final answer

+++ nn_predict.py
@@ -8,15 +8,15 @@ def relu(x):
 
 def softmax(x):
     # TODO: Implement the SoftMax function
-    x = np.asarray(x)
+    x = np.asarray(x, dtype=np.float32)
     if x.ndim == 1:
         x = x - np.max(x)
         e_x = np.exp(x)
-        return (e_x / np.sum(e_x)).astype(np.float32)
+        return e_x / np.sum(e_x)
     else:
         x = x - np.max(x, axis=-1, keepdims=True)
         e_x = np.exp(x)
-        return (e_x / np.sum(e_x, axis=-1, keepdims=True)).astype(np.float32)
+        return e_x / np.sum(e_x, axis=-1, keepdims=True)
 
 # === Flatten ===
 def flatten(x):
